Support arbitrary inner type in `NamedImplKey`.
Before this commit `NamedImplKey` could only represent the inner type as
`rust: &'a Ident`.  For example, it could represent `Vec<Foo>` but could
not represent `Vec<Box<Foo>>` where the inner type (`Box<Foo>` in our
example) is not a simple identifier.

After this commit `NamedImplKey` is refactored to support an arbitrary
inner type.  Note that this new ability is not actually used at this
point - follow-up commits will relax generic type argument restrictions
in `syntax/check.rs` one-by-one.

This commit is quite big, but it seems difficult to extract some changes
to smaller, separate commits, because all of the changes stem from the
refactoring of the `NamedImplKey`.  At a high-level this commit contains
the following changes:

1. `syntax/instantiate.rs`: Changing `pub inner: &'a Type` field of
   `NamedImplKey` to `pub inner: &'a Type`.  This is the main/root
   change in this commit.
2. `gen/src/write.rs`: supporting arbitrary inner types
   when writing C++ thunks exposing instantiations/monomorphizations of
   templates/generics supported by `cxx`.
    * This depends on `fn stringify_type` introduced in
      `gen/src/write.rs` in an earlier commit.
    * Handling arbitrary inner types *in general* means that we can
      delete `enum UniquePtr` which provided handling of two *specific*
      inner types.
3. `macro/src/expand.rs`: supporting arbitrary inner types
   when writing Rust thunks exposing instantiations/monomorphizations of
   templates/generics supported by `cxx`.
    * Using `#inner` instead of `#ident` may now (optionally) cover
      generic lifetime arguments.  This is why this commit also changes
      `macro/src/generics.rs`.
4. `syntax/mangle.rs`: supporting mangling of individual types.  This
   supports the high-level goal of allow arbitrary inner types, but it
   also helps to deduplicate mangling code that was somewhat duplicated
   in `macro/src/expand.rs` and `gen/src/write.rs`.
5. `syntax/types.rs`: Supporting arbitrary inner types in
    * `fn is_maybe_trivial`
    * `fn is_local` (this function supports an earlier refactoring
      that changed how `cxx` decides whether to provide an *implicit*
      impl of a given generic/template instantiation/monomorphization)