Emit error if `#[namespace = ...]` attribute of a method is ignored.
In the past methods in a different namespace from their receiver
wouldn't work.  This was fixed in a recent commit after which methods
always use the namespace of their receiver, effectively ignoring any
`#[namespace = ...]` attributes that may be apply either directly
or transitively (e.g. above - say on `extern "Rust"`) to the method.

This commit detects `#[namespace = ...]` attributes that apply
*directly* to methods and emits an error that such attributes have
no effect.